requirements.txt
python-dotenv
groq
pandas
numpy
chardet
